+++ crates/zng-task/src/channel/ipc_bytes.rs
@@ -5,6 +5,7 @@ use std::{
     fmt, fs,
     io::{self, Read, Write},
     marker::PhantomData,
+    mem::MaybeUninit,
     ops,
     path::{Path, PathBuf},
     pin::Pin,
@@ -98,10 +99,12 @@ impl IpcBytes {
         #[cfg(ipc)]
         if len <= Self::INLINE_MAX {
             Ok(IpcBytesMut {
+                len,
                 inner: IpcBytesMutInner::Heap(vec![0; len]),
             })
         } else if len <= Self::UNNAMED_MAX {
             Ok(IpcBytesMut {
+                len,
                 inner: IpcBytesMutInner::AnonMemMap(IpcSharedMemory::from_byte(0, len)),
             })
         } else {
@@ -111,6 +114,7 @@ impl IpcBytes {
         #[cfg(not(ipc))]
         {
             Ok(IpcBytesMut {
+                len,
                 inner: IpcBytesMutInner::Heap(vec![0; len]),
             })
         }
@@ -305,10 +309,12 @@ impl IpcBytes {
         #[cfg(ipc)]
         if len <= Self::INLINE_MAX {
             Ok(IpcBytesMut {
+                len,
                 inner: IpcBytesMutInner::Heap(vec![0; len]),
             })
         } else if len <= Self::UNNAMED_MAX {
             Ok(IpcBytesMut {
+                len,
                 inner: IpcBytesMutInner::AnonMemMap(IpcSharedMemory::from_byte(0, len)),
             })
         } else {
@@ -325,6 +331,7 @@ impl IpcBytes {
             // SAFETY: we hold write lock
             let map = unsafe { memmap2::MmapMut::map_mut(&file) }?;
             Ok(IpcBytesMut {
+                len,
                 inner: IpcBytesMutInner::MemMap {
                     name,
                     map,
@@ -335,6 +342,7 @@ impl IpcBytes {
         #[cfg(not(ipc))]
         {
             Ok(IpcBytesMut {
+                len,
                 inner: IpcBytesMutInner::Heap(vec![0; len]),
             })
         }
@@ -885,26 +893,31 @@ impl IpcBytesWriterBlocking {
         self.flush()?;
         #[cfg(ipc)]
         {
-            let inner = match self.memmap {
+            let (len, inner) = match self.memmap {
                 Some((name, write_handle)) => {
                     // SAFETY: we hold write lock
                     let map = unsafe { memmap2::MmapMut::map_mut(&write_handle) }?;
-                    IpcBytesMutInner::MemMap { name, map, write_handle }
+                    let len = map.len();
+                    (len, IpcBytesMutInner::MemMap { name, map, write_handle })
                 }
                 None => {
-                    if self.heap_buf.len() > IpcBytes::INLINE_MAX {
+                    let len = self.heap_buf.len();
+                    let i = if self.heap_buf.len() > IpcBytes::INLINE_MAX {
                         IpcBytesMutInner::AnonMemMap(IpcSharedMemory::from_bytes(&self.heap_buf))
                     } else {
                         IpcBytesMutInner::Heap(self.heap_buf)
-                    }
+                    };
+                    (len, i)
                 }
             };
-            Ok(IpcBytesMut { inner })
+            Ok(IpcBytesMut { len, inner })
         }
         #[cfg(not(ipc))]
         {
-            let inner = IpcBytesMutInner::Heap(self.heap_buf.into_inner());
-            Ok(IpcBytesMut { inner })
+            let heap_buf = self.heap_buf.into_inner();
+            let len = heap_buf.len();
+            let inner = IpcBytesMutInner::Heap(heap_buf);
+            Ok(IpcBytesMut { len, inner })
         }
     }
 
@@ -1010,31 +1023,34 @@ enum IpcBytesMutInner {
 /// Use [`IpcBytes::new_mut`] or [`IpcBytes::new_mut_blocking`] to allocate.
 pub struct IpcBytesMut {
     inner: IpcBytesMutInner,
+    len: usize,
 }
 impl ops::Deref for IpcBytesMut {
     type Target = [u8];
 
     fn deref(&self) -> &Self::Target {
+        let len = self.len;
         match &self.inner {
-            IpcBytesMutInner::Heap(v) => &v[..],
+            IpcBytesMutInner::Heap(v) => &v[..len],
             #[cfg(ipc)]
-            IpcBytesMutInner::AnonMemMap(m) => &m[..],
+            IpcBytesMutInner::AnonMemMap(m) => &m[..len],
             #[cfg(ipc)]
-            IpcBytesMutInner::MemMap { map, .. } => &map[..],
+            IpcBytesMutInner::MemMap { map, .. } => &map[..len],
         }
     }
 }
 impl ops::DerefMut for IpcBytesMut {
     fn deref_mut(&mut self) -> &mut Self::Target {
+        let len = self.len;
         match &mut self.inner {
-            IpcBytesMutInner::Heap(v) => &mut v[..],
+            IpcBytesMutInner::Heap(v) => &mut v[..len],
             #[cfg(ipc)]
             IpcBytesMutInner::AnonMemMap(m) => {
                 // SAFETY: we are the only reference to the map
                 unsafe { m.deref_mut() }
             }
             #[cfg(ipc)]
-            IpcBytesMutInner::MemMap { map, .. } => &mut map[..],
+            IpcBytesMutInner::MemMap { map, .. } => &mut map[..len],
         }
     }
 }
@@ -1049,6 +1065,7 @@ impl IpcBytesMut {
         #[cfg(ipc)]
         if buf.len() <= IpcBytes::INLINE_MAX {
             Ok(Self {
+                len: buf.len(),
                 inner: IpcBytesMutInner::Heap(buf),
             })
         } else {
@@ -1062,6 +1079,7 @@ impl IpcBytesMut {
         #[cfg(not(ipc))]
         {
             Ok(Self {
+                len: buf.len(),
                 inner: IpcBytesMutInner::Heap(buf),
             })
         }
@@ -1069,23 +1087,46 @@ impl IpcBytesMut {
 
     /// Convert to immutable shareable [`IpcBytes`].
     pub async fn finish(mut self) -> io::Result<IpcBytes> {
+        let len = self.len;
         let data = match std::mem::replace(&mut self.inner, IpcBytesMutInner::Heap(vec![])) {
-            IpcBytesMutInner::Heap(v) => IpcBytesData::Heap(v),
+            IpcBytesMutInner::Heap(mut v) => {
+                v.truncate(len);
+                v.shrink_to_fit();
+                IpcBytesData::Heap(v)
+            }
             #[cfg(ipc)]
-            IpcBytesMutInner::AnonMemMap(m) => IpcBytesData::AnonMemMap(m),
+            IpcBytesMutInner::AnonMemMap(m) => {
+                if len < IpcBytes::INLINE_MAX {
+                    IpcBytesData::Heap(m[..len].to_vec())
+                } else if len < m.len() {
+                    IpcBytesData::AnonMemMap(IpcSharedMemory::from_bytes(&m[..len]))
+                } else {
+                    IpcBytesData::AnonMemMap(m)
+                }
+            }
             #[cfg(ipc)]
             IpcBytesMutInner::MemMap { name, map, write_handle } => {
-                blocking::unblock(move || Self::finish_memmap(name, map, write_handle)).await?
+                let len = self.len;
+                blocking::unblock(move || Self::finish_memmap(name, map, write_handle, len)).await?
             }
         };
         Ok(IpcBytes(Arc::new(data)))
     }
 
     #[cfg(ipc)]
-    fn finish_memmap(name: PathBuf, map: memmap2::MmapMut, write_handle: fs::File) -> Result<IpcBytesData, io::Error> {
-        let len = map.len();
+    fn finish_memmap(name: PathBuf, map: memmap2::MmapMut, write_handle: fs::File, len: usize) -> Result<IpcBytesData, io::Error> {
+        let alloc_len = map.len();
+        if alloc_len != len {
+            write_handle.set_len(len as u64)?;
+        }
         write_handle.unlock()?;
-        let map = map.make_read_only()?;
+        let map = if alloc_len != len {
+            drop(map);
+            // SAFETY: we have write access to the file still
+            unsafe { memmap2::Mmap::map(&write_handle) }?
+        } else {
+            map.make_read_only()?
+        };
         let mut permissions = write_handle.metadata()?.permissions();
         permissions.set_readonly(true);
         #[cfg(unix)]
@@ -1112,6 +1153,7 @@ impl IpcBytesMut {
         #[cfg(ipc)]
         if buf.len() <= IpcBytes::INLINE_MAX {
             Ok(Self {
+                len: buf.len(),
                 inner: IpcBytesMutInner::Heap(buf),
             })
         } else {
@@ -1122,6 +1164,7 @@ impl IpcBytesMut {
         #[cfg(not(ipc))]
         {
             Ok(Self {
+                len: self.len,
                 inner: IpcBytesMutInner::Heap(buf),
             })
         }
@@ -1129,12 +1172,24 @@ impl IpcBytesMut {
 
     /// Convert to immutable shareable [`IpcBytes`].
     pub fn finish_blocking(mut self) -> io::Result<IpcBytes> {
+        let len = self.len;
         let data = match std::mem::replace(&mut self.inner, IpcBytesMutInner::Heap(vec![])) {
-            IpcBytesMutInner::Heap(v) => IpcBytesData::Heap(v),
+            IpcBytesMutInner::Heap(mut v) => {
+                v.truncate(len);
+                IpcBytesData::Heap(v)
+            }
             #[cfg(ipc)]
-            IpcBytesMutInner::AnonMemMap(m) => IpcBytesData::AnonMemMap(m),
+            IpcBytesMutInner::AnonMemMap(m) => {
+                if len < IpcBytes::INLINE_MAX {
+                    IpcBytesData::Heap(m[..len].to_vec())
+                } else if len < m.len() {
+                    IpcBytesData::AnonMemMap(IpcSharedMemory::from_bytes(&m[..len]))
+                } else {
+                    IpcBytesData::AnonMemMap(m)
+                }
+            }
             #[cfg(ipc)]
-            IpcBytesMutInner::MemMap { name, map, write_handle } => Self::finish_memmap(name, map, write_handle)?,
+            IpcBytesMutInner::MemMap { name, map, write_handle } => Self::finish_memmap(name, map, write_handle, len)?,
         };
         Ok(IpcBytes(Arc::new(data)))
     }
@@ -1275,5 +1330,185 @@ impl IpcBytes {
     /// [bytemuck docs]: https://docs.rs/bytemuck/1.24.0/bytemuck/fn.try_cast_slice.html
     pub fn cast_deref<T: bytemuck::AnyBitPattern>(&self) -> &[T] {
         bytemuck::cast_slice(self)
+    }
+}
+
+impl IpcBytesMut {
+    /// Shorten the bytes length.
+    ///
+    /// If `new_len` is greater or equal to current length does nothing.
+    ///
+    /// Note that this does not affect memory allocation, the extra bytes are only dropped on finish.
+    pub fn truncate(&mut self, new_len: usize) {
+        self.len = self.len.min(new_len);
+    }
+
+    /// Convert chunks of length `L0` to chunks of length `L1` where `L1 <= L0`.
+    ///
+    /// Reuses the same allocation for the new data, replacing in place. Truncates the final buffer to the new length.
+    ///
+    /// Note that no memory is released by this call as truncated is lazy and applied on finish.
+    ///
+    /// # Panics
+    ///
+    /// Panics if `L1 > L0` or if bytes length is not multiple of `L0`.
+    pub fn reduce_in_place<const L0: usize, const L1: usize>(&mut self, mut reduce: impl FnMut([u8; L0]) -> [u8; L1]) {
+        assert!(L1 <= L0);
+        let len = self.len();
+        if len == 0 {
+            return;
+        }
+        assert!(len.is_multiple_of(L0), "length must be multiple of L0");
+
+        let ptr = self.as_mut_ptr();
+        let mut write = 0usize;
+        let mut read = 0usize;
+
+        // SAFETY: pointers stay inside slice, in_chunk and out_chunk copy never overlaps with slice.
+        unsafe {
+            while read < len {
+                let mut in_chunk = MaybeUninit::<[u8; L0]>::uninit();
+                std::ptr::copy_nonoverlapping(ptr.add(read), (*in_chunk.as_mut_ptr()).as_mut_ptr(), L0);
+                read += L0;
+
+                let out_chunk = reduce(in_chunk.assume_init());
+
+                std::ptr::copy_nonoverlapping(out_chunk.as_ptr(), ptr.add(write), L1);
+                write += L1;
+            }
+        }
+
+        self.truncate(write);
+    }
+
+    /// Convert chunks of `in_chunk_len` to chunks of `out_chunk_buf.len()` where `out_chunk_buf.len() <= in_chunk_len`.
+    ///
+    /// Reuses the same allocation for the new data, replacing in place. Truncates the final buffer to the new length.
+    ///
+    /// Note that no memory is released by this call as truncated is lazy and applied on finish.
+    ///
+    /// # Panics
+    ///
+    /// Panics if `out_chunk_buf.len() > in_chunk_len` or if bytes length is not multiple of `in_chunk_len`.
+    pub fn reduce_in_place_dyn(&mut self, in_chunk_len: usize, out_chunk_buf: &mut [u8], mut reduce: impl FnMut(&[u8], &mut [u8])) {
+        assert!(out_chunk_buf.len() < in_chunk_len);
+        let len = self.len();
+        if len == 0 {
+            return;
+        }
+        assert!(len.is_multiple_of(in_chunk_len), "length must be multiple of in_chunk_len");
+
+        let ptr = self.as_mut_ptr();
+        let mut write = 0usize;
+        let mut read = 0usize;
+
+        // SAFETY: pointers stay inside slice, in_chunk and out_chunk copy never overlaps with slice.
+        unsafe {
+            while read < len {
+                reduce(std::slice::from_raw_parts(ptr.add(read), in_chunk_len), &mut *out_chunk_buf);
+                read += in_chunk_len;
+
+                std::ptr::copy_nonoverlapping(out_chunk_buf.as_ptr(), ptr.add(write), out_chunk_buf.len());
+                write += out_chunk_buf.len();
+            }
+        }
+
+        self.truncate(write);
+    }
+
+    /// Convert chunks of length `L0` to chunks of length `L1` where `size_of::<T1>() * L1 <= size_of::<T0>() * L0`.
+    ///
+    /// Reuses the same allocation for the new data, replacing in place. Truncates the final buffer to the new length.
+    ///
+    /// Note that no memory is released by this call as truncated is lazy and applied on finish.
+    ///
+    /// # Panics
+    ///
+    /// Panics if `size_of::<T1>() * L1 > size_of::<T0>() * L0` or if bytes length is not multiple of `size_of::<T0>() * L0`.
+    pub fn cast_reduce_in_place<T0, const L0: usize, T1, const L1: usize>(&mut self, mut reduce: impl FnMut([T0; L0]) -> [T1; L1])
+    where
+        T0: bytemuck::AnyBitPattern,
+    {
+        let l0 = std::mem::size_of::<T0>() * L0;
+        let l1 = std::mem::size_of::<T1>() * L1;
+
+        assert!(l1 <= l0);
+        let len = self.len();
+        if len == 0 {
+            return;
+        }
+        assert!(len.is_multiple_of(l0), "length must be multiple of size_of::<T0>() * L0");
+
+        let ptr = self.as_mut_ptr();
+        let mut write = 0usize;
+        let mut read = 0usize;
+
+        // SAFETY:
+        // Pointers stay inside slice, in_chunk and out_chunk copy never overlaps with slice.
+        // Reading [T0; L0] from raw bytes is safe because T0: AnyBitPattern
+        unsafe {
+            while read < len {
+                let mut in_chunk = MaybeUninit::<[T0; L0]>::uninit();
+                std::ptr::copy_nonoverlapping(ptr.add(read), (*in_chunk.as_mut_ptr()).as_mut_ptr() as _, l0);
+                read += l0;
+
+                let out_chunk = reduce(in_chunk.assume_init());
+
+                std::ptr::copy_nonoverlapping(out_chunk.as_ptr() as _, ptr.add(write), l1);
+                write += l1;
+            }
+        }
+
+        self.truncate(write);
+    }
+
+    /// Convert chunks of `size_of::<T0>() * in_chunk_len` to chunks of `size_of::<T1>() * out_chunk_buf.len()`
+    /// where `size_of::<T1>() * out_chunk_buf.len() <= size_of::<T0>() * in_chunk_len`.
+    ///
+    /// Reuses the same allocation for the new data, replacing in place. Truncates the final buffer to the new length.
+    ///
+    /// Note that no memory is released by this call as truncated is lazy and applied on finish.
+    ///
+    /// # Panics
+    ///
+    /// Panics if `size_of::<T1>() * out_chunk_buf.len() > size_of::<T0>() * in_chunk_len` or if bytes
+    /// length is not multiple of `size_of::<T0>() * in_chunk_len`.
+    pub fn cast_reduce_in_place_dyn<T0, T1>(
+        &mut self,
+        in_chunk_len: usize,
+        out_chunk_buf: &mut [T1],
+        mut reduce: impl FnMut(&[T0], &mut [T1]),
+    ) where
+        T0: bytemuck::AnyBitPattern,
+    {
+        let l0 = std::mem::size_of::<T0>() * in_chunk_len;
+        let l1 = std::mem::size_of_val(out_chunk_buf);
+
+        assert!(l1 <= l0);
+        let len = self.len();
+        if len == 0 {
+            return;
+        }
+        assert!(len.is_multiple_of(l0), "length must be multiple of size_of::<T0>() * L0");
+
+        let ptr = self.as_mut_ptr();
+        let mut write = 0usize;
+        let mut read = 0usize;
+
+        // SAFETY: pointers stay inside slice, in_chunk and out_chunk copy never overlaps with slice.
+        unsafe {
+            while read < len {
+                reduce(
+                    bytemuck::cast_slice(std::slice::from_raw_parts(ptr.add(read), l0)),
+                    &mut *out_chunk_buf,
+                );
+                read += l0;
+
+                std::ptr::copy_nonoverlapping(out_chunk_buf.as_ptr() as _, ptr.add(write), l1);
+                write += l1;
+            }
+        }
+
+        self.truncate(write);
     }
 }